requirements.txt
django
djangorestframework
environs
selenium
bs4
pyperclip
openpyxl